# config/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Task Manager",
    "description": "REST API for Task Manager",
    "version": "0.0.1"
  },
...
  "paths": {
    "/auth/test": {
      "get": {
        "summary": "Проверка работы сервера",
        "responses": {
          "200": {
            "description": "Работает",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Не работает",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/registration": {
      "post": {
        "summary": "Создать пользователя",
        "requestBody": {
          "$ref": "#/components/requestBodies/Auth"
        },
        "responses": {
          "201": {
            "description": "Пользователь создан",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "text": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Ошибка",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "text": {
                      "type": "string"
                    },
                    "error": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Авторизация по логину и паролю",
        "requestBody": {
          "$ref": "#/components/requestBodies/Auth"
        },
        "responses": {
          "201": {
            "description": "Успешная авторизация",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Ошибка авторизации, проверьте логин и пароль",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Ошибка",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/me": {
      "post": {
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "Проверка токена",
        "responses": {
          "200": {
            "description": "Токен рабочий",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "userId": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Нет доступа",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Ошибка",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "Authorization": {
        "type": "apiKey",
        "name": "authorization",
        "in": "header"
      }
    },
    "requestBodies": {
      "Auth": {
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "format": "email"
                },
                "password": {
                  "type": "string"
                }
              },
              "required": [
                "email",
                "password"
              ]
            }
          }
        }
      },
...
    },
...
  }
}